backend/mazerunner/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1.1
django-nested-admin==3.3.2
djangorestframework==3.11.1
filelock==3.0.12
Markdown==3.2.2
python-monkey-business==1.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
